update: rename models name
